+++ nv/plugin_multiple_cursors.py
@@ -1,0 +1,46 @@
+# Copyright (C) 2018 The NeoVintageous Team (NeoVintageous).
+#
+# This file is part of NeoVintageous.
+#
+# NeoVintageous is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# NeoVintageous is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with NeoVintageous.  If not, see <https://www.gnu.org/licenses/>.
+
+# A port of https://github.com/terryma/vim-multiple-cursors
+# Note that some of the command definitions are in nv/cmd_defs.
+
+from NeoVintageous.nv.plugin import register
+from NeoVintageous.nv.vi.cmd_base import ViOperatorDef
+from NeoVintageous.nv.vim import NORMAL
+from NeoVintageous.nv.vim import VISUAL
+
+
+__all__ = ()
+
+
+@register(seq='<C-n>', modes=(NORMAL, VISUAL))
+class _multiple_cursors_enter(ViOperatorDef):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.updates_xpos = True
+        self.scroll_into_view = True
+
+    def is_enabled(self, settings):
+        return settings.get('vintageous_enable_multiple_cursors')
+
+    def translate(self, state):
+        return {
+            'action': '_enter_select_mode',
+            'action_args': {
+                'mode': state.mode
+            }
+        }