+++ plugin.py
@@ -63,6 +63,7 @@ try:
     # Plugins.
     from NeoVintageous.nv.plugin_abolish import *  # noqa: F401,F403
     from NeoVintageous.nv.plugin_commentary import *  # noqa: F401,F403
+    from NeoVintageous.nv.plugin_multiple_cursors import *  # noqa: F401,F403
     from NeoVintageous.nv.plugin_surround import *  # noqa: F401,F403
     from NeoVintageous.nv.plugin_unimpaired import *  # noqa: F401,F403