+++ tests/functional/test__esc.py
@@ -49,6 +49,13 @@ class Test_esc(unittest.FunctionalTestCase):
         self.eq('r_f|iz|z\nb|uz|z\n', 'b_<esc>', 'n_fizz\nb|uzz\n')
         self.assertStatusLineIsBlank()
 
+    def test_s(self):
+        self.eq('f|iz|z\nb|uz|z\n', 's_<esc>', 'n_f|izz\nbuzz\n')
+        self.settings().set('vintageous_multi_cursor_exit_from_visual_mode', False)
+        self.eq('f|iz|z\nb|uz|z\n', 's_<esc>', 'n_f|izz\nb|uzz\n')
+        self.settings().set('vintageous_multi_cursor_exit_from_visual_mode', True)
+        self.eq('f|iz|z\nb|uz|z\n', 's_<esc>', 'n_f|izz\nbuzz\n')
+
     def test_N_esc(self):
         self.eq('fi|zzb|uzz', '<esc>', 'N_fi|zzb|uzz')
         self.assertStatusLineIsBlank()